+++ src/python/pants/backend/go/module_integration_test.py
@@ -4,7 +4,7 @@ import textwrap
 
 import pytest
 
-from pants.backend.go import module, sdk
+from pants.backend.go import module
 from pants.backend.go.module import (
     DownloadedExternalModule,
     DownloadExternalModuleRequest,
@@ -12,6 +12,7 @@ from pants.backend.go.module import (
     ResolveGoModuleRequest,
 )
 from pants.backend.go.target_types import GoModule, GoPackage
+from pants.backend.go.util_rules import sdk
 from pants.build_graph.address import Address
 from pants.core.util_rules import external_tool, source_files
 from pants.engine import fs
