+++ src/python/pants/backend/go/tailor_test.py
@@ -4,7 +4,7 @@ import textwrap
 
 import pytest
 
-from pants.backend.go import module, sdk, target_type_rules
+from pants.backend.go import module, target_type_rules
 from pants.backend.go.tailor import (
     PutativeGoExternalModuleTargetsRequest,
     PutativeGoModuleTargetsRequest,
@@ -12,6 +12,7 @@ from pants.backend.go.tailor import (
 )
 from pants.backend.go.tailor import rules as go_tailor_rules
 from pants.backend.go.target_types import GoExternalPackageTarget, GoModule, GoPackage
+from pants.backend.go.util_rules import sdk
 from pants.core.goals.tailor import (
     AllOwnedSources,
     PutativeTarget,
